criacao modulo planos

+++ app/Http/Controllers/Admin/PlanController.php
@@ -48,11 +48,19 @@ class PlanController extends Controller
     }
 
     public function destroy ($id) {
-        $plan = $this->repository->where('id', $id)->first();
+        $plan = $this->repository
+//                        ->with('details')
+                        ->where('id', $id)
+                        ->first();
 
         if (!$plan) {
             return redirect()->back();
         }
+
+        if ($plan->details->count() > 0) {
+            return redirect()->back()->with('error', 'Existem detalhes vinculado a esse plano, portanto não é possivel deletar-lo');
+        }
+
         $plan->delete();
         return redirect()->route('plans.index');
     }
